Revert broken cleanup code. The else was accidentally attached to the try block, causing all cores to be written to disk followed by a 400 error being send and nothing put on the rabbit queue.

+++ submit_core.wsgi
@@ -68,19 +68,21 @@ def wsgi_handler(environ, start_response):
             if not addr_sig:
                 start_response('400 Bad Request', [])
                 return ['']
+            copied = False
             with open(path, 'w') as fp:
                 try:
                     shutil.copyfileobj(environ['wsgi.input'], fp, 512)
+                    copied = True
                 except IOError, e:
-                    try:
-                        os.remove(path)
-                    except OSError:
-                        pass
                     if e.message != 'request data read error':
                         raise
-                else:
-                    start_response('400 Bad Request', [])
-                    return ['']
+            if not copied:
+                try:
+                    os.remove(path)
+                except OSError:
+                    pass
+                start_response('400 Bad Request', [])
+                return ['']
 
             os.chmod(path, 0o666)
             channel.queue_declare(queue=queue, durable=True, auto_delete=False)
